+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # v0.1.1
+- Handle broken UTF8 in `show` with `-f` and `-e` flags.
+- Allow restart of `Killed` processes
+
+# v0.1.1
 
 - Replace prettytables-rs with comfy-table
 - Replace termion with crossterm